+++ kubejs/startup_scripts/create/constants.js
@@ -449,7 +449,6 @@ global.CREATE_DISABLED_ITEMS = [
     'create:refined_radiance_casing', 
     'create:shadow_steel_casing', 
     'create:wooden_bracket', 
-    'create:blaze_burner',
     'create:sand_paper',
     'create:red_sand_paper',
     'create:empty_blaze_burner',