+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Moved MV superconductor to early HV instead of late HV (Pyritie).
 - Added some Italy and Ukraine translation (MetenBouldry).
 - Сreate some rich coal recipes in coke oven and pyro oven (Exzept1on).
+- Add blaze burner, fix alternator (Pyritie).
 ### New mods
 - Inventory Tweaks Refoxed 1.1
 - EveryCompat